Refactor quart Routes
