+++ app.py
@@ -1,9 +1,20 @@
+from src.python.route.quartMainRoute import mainRoute
+from src.python.route.quartTrackerRoute import trackerRoute
+from src.python.route.quartEmailRoute import emailRoute
 from src.python.route.quartWebSocketRoute import webSocketRoute
-from src.python.route.quartServerRoute import serverRoute
-from src.python.config.appConfig import app, quartDebug, quartEnv, quartHost, quartPort
+from src.python.config.appConfig import (
+    app,
+    quartDebug,
+    quartEnv,
+    quartHost,
+    quartPort,
+)
 
 
-_ = [app.register_blueprint(route) for route in [webSocketRoute, serverRoute]]
+_ = [
+    app.register_blueprint(route)
+    for route in [mainRoute, trackerRoute, emailRoute, webSocketRoute]
+]
 
 
 if __name__ == "__main__":
